+++ docs/project/changelog.md
@@ -183,6 +183,8 @@ substitutions:
   some completion issues (see
   [#821](https://github.com/pyodide/pyodide/issues/821) and
   [#1160](https://github.com/pyodide/pyodide/issues/1160)
+- {{ Enhancement }} Support top-level await in the console
+  [#1459](https://github.com/pyodide/pyodide/issues/1459)
 
 ### Packages